style: move habits stack to separate folder
